Add Lrmi base class. Move Finna specific Qdc fucntions to trait.
