+++ src/RecordManager/Finna/Record/Qdc.php
@@ -27,8 +27,6 @@
  */
 namespace RecordManager\Finna\Record;
 
-use RecordManager\Base\Utils\MetadataUtils;
-
 /**
  * Qdc record class
  *
@@ -42,192 +40,5 @@ use RecordManager\Base\Utils\MetadataUtils;
  */
 class Qdc extends \RecordManager\Base\Record\Qdc
 {
-    /**
-     * Return fields to be indexed in Solr
-     *
-     * @param \RecordManager\Base\Database\Database $db Database connection. Omit to
-     *                                                  avoid database lookups for
-     *                                                  related records.
-     *
-     * @return array
-     */
-    public function toSolrArray(\RecordManager\Base\Database\Database $db = null)
-    {
-        $data = parent::toSolrArray($db);
-
-        // Nonstandard author fields
-        $authors = $this->getValues('author');
-        if ($authors) {
-            $data['author'] = array_shift($authors);
-            if (isset($data['author2'])) {
-                $data['author2'] = array_merge($authors, $data['author2']);
-            } else {
-                $data['author2'] = $authors;
-            }
-        }
-
-        if (isset($data['publishDate'])) {
-            $data['main_date_str']
-                = MetadataUtils::extractYear($data['publishDate']);
-            $data['main_date'] = $this->validateDate(
-                $this->getPublicationYear() . '-01-01T00:00:00Z'
-            );
-        }
-
-        if ($range = $this->getPublicationDateRange()) {
-            $data['search_daterange_mv'][] = $data['publication_daterange']
-                = MetadataUtils::dateRangeToStr($range);
-        }
-
-        foreach ($this->doc->relation as $relation) {
-            $url = trim((string)$relation);
-            // Ignore too long fields. Require at least one dot surrounded by valid
-            // characters or a familiar scheme
-            if (strlen($url) > 4096
-                || (!preg_match('/^[A-Za-z0-9]\.[A-Za-z0-9]$/', $url)
-                && !preg_match('/^(http|ftp)s?:\/\//', $url))
-            ) {
-                continue;
-            }
-            $link = [
-                'url' => $url,
-                'text' => '',
-                'source' => $this->source
-            ];
-            $data['online_boolean'] = true;
-            $data['online_str_mv'] = $this->source;
-            // Mark everything free until we know better
-            $data['free_online_boolean'] = true;
-            $data['free_online_str_mv'] = $this->source;
-            $data['online_urls_str_mv'][] = json_encode($link);
-        }
-
-        foreach ($this->doc->file as $file) {
-            $url = (string)$file->attributes()->href
-                ? trim((string)$file->attributes()->href)
-                : trim((string)$file);
-            $link = [
-                'url' => $url,
-                'text' => trim((string)$file->attributes()->name),
-                'source' => $this->source
-            ];
-            $data['online_boolean'] = true;
-            $data['online_str_mv'] = $this->source;
-            $data['free_online_boolean'] = true;
-            // Mark everything free until we know better
-            $data['free_online_str_mv'] = $this->source;
-            $data['online_urls_str_mv'][] = json_encode($link);
-            if (strcasecmp($file->attributes()->bundle, 'THUMBNAIL') == 0
-                && !isset($data['thumbnail'])
-            ) {
-                $data['thumbnail'] = $url;
-            }
-        }
-
-        foreach ($this->doc->coverage as $coverage) {
-            $attrs = $coverage->attributes();
-            if ($attrs->type == 'geocoding') {
-                $match = preg_match(
-                    '/([\d\.]+)\s*,\s*([\d\.]+)/', trim((string)$coverage), $matches
-                );
-                if ($match) {
-                    if ($attrs->format == 'lon,lat') {
-                        $lon = $matches[1];
-                        $lat = $matches[2];
-                    } else {
-                        $lat = $matches[1];
-                        $lon = $matches[2];
-                    }
-                    $data['location_geo'][] = "POINT($lon $lat)";
-                }
-            }
-        }
-        if (!empty($data['location_geo'])) {
-            $data['center_coords']
-                = MetadataUtils::getCenterCoordinates($data['location_geo']);
-        }
-
-        // Usage rights
-        if ($rights = $this->getUsageRights()) {
-            $data['usage_rights_str_mv'] = $rights;
-        }
-
-        $data['source_str_mv'] = $this->source;
-        $data['datasource_str_mv'] = $this->source;
-
-        $data['author_facet'] = array_merge(
-            isset($data['author']) ? (array)$data['author'] : [],
-            isset($data['author2']) ? (array)$data['author2'] : [],
-            isset($data['author_corporate']) ? (array)$data['author_corporate'] : []
-        );
-
-        $data['format_ext_str_mv'] = $data['format'];
-
-        return $data;
-    }
-
-    /**
-     * Return publication year/date range
-     *
-     * @return array|null
-     */
-    protected function getPublicationDateRange()
-    {
-        $year = $this->getPublicationYear();
-        if ($year) {
-            $startDate = "$year-01-01T00:00:00Z";
-            $endDate = "$year-12-31T23:59:59Z";
-            return [$startDate, $endDate];
-        }
-        return null;
-    }
-
-    /**
-     * Return usage rights if any
-     *
-     * @return array ['restricted'] or more specific id's if defined for the record
-     */
-    protected function getUsageRights()
-    {
-        if (!isset($this->doc->rights)) {
-            return ['restricted'];
-        }
-        $result = [];
-        foreach ($this->doc->rights as $rights) {
-            $result[] = trim((string)$rights);
-        }
-        return $result;
-    }
-
-    /**
-     * Return URLs associated with object
-     *
-     * @return array
-     */
-    protected function getUrls()
-    {
-        $urls = parent::getUrls();
-
-        if ($this->doc->permaddress) {
-            $urls[] = trim((string)$this->doc->permaddress[0]);
-        }
-
-        foreach ($this->getValues('identifier') as $identifier) {
-            $res = preg_match(
-                '/^(URN:NBN:fi:|URN:ISBN:978-?951|URN:ISBN:951)/i', $identifier
-            );
-            if ($res) {
-                if (!empty($urls)) {
-                    // Check that the identifier is not already listed
-                    foreach ($urls as $url) {
-                        if (stristr($url, $identifier) !== false) {
-                            continue 2;
-                        }
-                    }
-                }
-                $urls[] = "http://urn.fi/$identifier";
-            }
-        }
-        return $urls;
-    }
+    use QdcRecordTrait;
 }
